+++ lib/sudoku_rb.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "sudoku_rb/version"
-
-module SudokuRb
-  class Error < StandardError; end
-  # Your code goes here...
-end